+++ Core/AuthorModel.cs
@@ -8,7 +8,19 @@ namespace Core
     /// </summary>
     public class AuthorModel
     {
-        protected DbPublicationEntities context = new DbPublicationEntities();
+        /// <summary>
+        /// Uchovává databázový kontext.
+        /// </summary>
+        protected DbPublicationEntities context;
+
+        /// <summary>
+        /// Vytvoří instanci správce.
+        /// </summary>
+        /// <param name="context">databázový kontext</param>
+        public AuthorModel(DbPublicationEntities context)
+        {
+            this.context = context;
+        }
 
         /// <summary>
         /// Vrátí seznam všech uložených autorů.
@@ -28,23 +40,22 @@ namespace Core
         }
 
         /// <summary>
-        /// Vrátí seznam autorů publikace se zadaným ID.
+        /// Vrátí uloženého autora podle ID.
         /// </summary>
-        /// <param name="publicationId">ID publikace</param>
-        /// <returns>seznam autorů publikace</returns>
-        /*public List<Author> GetAuthorsByPublication(int publicationId)
+        /// <param name="id">ID autora</param>
+        /// <returns>autor</returns>
+        public Author GetAuthorById(int id)
         {
-            Publication publication = context.Publication.Find(publicationId);
+            Author author = context.Author.Find(id);
 
-            // výběr autorů podle ID publikace
-            var authors = from a in context.Author
-                            where a.Publication == publication
-                            orderby a.Name
-                            select a;
+            if (author == null)
+            {
+                throw new AuthorException(string.Format("Autor s id {0} neexistuje.", id));
+            }
 
-            return authors.ToList();
-        }*/
-
+            return author;
+        }
+        
         /// <summary>
         /// Uloží údaje nového autora pro pozdější použití.
         /// </summary>
@@ -56,52 +67,18 @@ namespace Core
         }
 
         /// <summary>
-        /// Přidá nového autora do seznamu autorů publikace a uloží jeho údaje
-        /// pro pozdější použití.
-        /// </summary>
-        /// <param name="publicationId">ID publikace</param>
-        /// <param name="author">údaje o autorovi</param>
-        /*public void CreateAuthorOfPublication(Publication publication, Author author)
-        {
-            author.Publication.Add(publication);
-            publication.Author.Add(author);
-            context.Author.Add(author);
-            context.SaveChanges();
-        }*/
-
-        /// <summary>
-        /// Přidá existujícího autora do seznamu autorů publikace.
-        /// </summary>
-        /// <param name="publicationId">ID publikace</param>
-        /// <param name="author">údaje o autorovi</param>
-        /*public void AddAuthorToPublication(Publication publication, int authorId)
-        {
-            Author author = context.Author.Find(authorId);
-            author.Publication.Add(publication);
-            publication.Author.Add(author);
-            context.SaveChanges();
-        }*/
-
-        /// <summary>
-        /// Odebere autora ze seznamu autorů publikace.
-        /// </summary>
-        /// <param name="publicationId">ID publikace</param>
-        /// <param name="authorId">údaje o autorovi</param>
-        /*public void RemoveAuthorFromPublication(Publication publication, int authorId)
-        {
-            Author author = context.Author.Find(authorId);
-            publication.Author.Remove(author);
-            author.Publication.Remove(publication);
-            context.SaveChanges();
-        }*/
-
-        /// <summary>
         /// Odstraní uložený záznam o autorovi. Lze použít pouze na autora bez publikací.
         /// </summary>
-        /// <param name="authorId"></param>
-        public void DeleteAuthor(int authorId)
+        /// <param name="id"></param>
+        public void DeleteAuthor(int id)
         {
-            Author author = context.Author.Find(authorId);
+            Author author = context.Author.Find(id);
+
+            if (author == null)
+            {
+                throw new AuthorException(string.Format("Autor s id {0} neexistuje.", id));
+            }
+
             context.Author.Remove(author);
             context.SaveChanges();
         }
